requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ── UI ─────────────────────────────────────────────────────────
streamlit>=1.40.0

# ── Vector search ──────────────────────────────────────────────
# faiss-cpu for most environments
# Use faiss-gpu if a CUDA-compatible GPU is available
faiss-cpu>=1.9.0

# ── Embeddings (neural backend) ────────────────────────────────
# Required only when USE_NEURAL_EMBEDDINGS=true in .env
# Downloads paraphrase-multilingual-MiniLM-L12-v2 (~420 MB) on first run
sentence-transformers>=2.2.0

# ── Language detection ─────────────────────────────────────────
langdetect>=1.0.9

# ── LLM provider SDKs ──────────────────────────────────────────
# Install the SDK for whichever provider you use (or both).
# Anthropic — set ANTHROPIC_API_KEY in .env
anthropic>=0.40.0
# Groq (free tier) — set GROQ_API_KEY in .env, swap llm_service.py
groq>=0.9.0

# ── Data and utilities ─────────────────────────────────────────
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0

# ── Testing ────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0

# ── Notebooks (optional) ───────────────────────────────────────
# Only needed if running notebooks/synthetic_data_generation.ipynb
# jupyter>=1.0.0
# ipykernel>=6.0.0
